# openapi/stood.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stood platform API",
    "version": "1.1.0",
    "summary": "Milestone payments held on PayPal (sandbox) and released only on evidence",
    "description": "Every request carries the platform key (Authorization: Bearer) and a Stood-Signature: t=<unix seconds>,v2=<hex HMAC-SHA256 over the canonical request>, within five minutes. Bodies are at most 64 KiB. Stood only talks to the PayPal sandbox.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://stood-api.onrender.com/v1",
      "description": "Hosted (PayPal sandbox)"
    }
  ],
  "paths": {
    "/allowances": {
      "post": {
        "operationId": "createAllowance",
        "summary": "Draft an allowance: a payment cap split into milestones, each with an evidence profile",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Same key and body return the same response; a different body is 409",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllowanceDraft"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Allowance"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or version conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Body over 64 KiB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/allowances/{id}": {
      "get": {
        "operationId": "getAllowance",
        "summary": "Read an allowance this platform owns",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Allowance"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/allowances/{id}/mandate": {
      "post": {
        "operationId": "requestMandate",
        "summary": "Start saved-PayPal signing for the allowance's current terms (send {})",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Same key and body return the same response; a different body is 409",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted (recorded, not yet done)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandate"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or version conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/allowances/{id}/mandate/{key}": {
      "get": {
        "operationId": "getMandate",
        "summary": "Read signing status and, while awaiting the buyer, the approval link",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mandate"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/baselines": {
      "post": {
        "operationId": "requestBaseline",
        "summary": "Ask Stood to run one milestone's frozen tests on its base commit (a red baseline proves new work is needed)",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Same key and body return the same response; a different body is 409",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeTerms"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted (recorded, not yet done)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Baseline"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or version conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Body over 64 KiB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/baselines/{id}": {
      "get": {
        "operationId": "getBaseline",
        "summary": "Read a baseline: queued, or each frozen test as it ran on the base commit",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Baseline"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/tranches/{id}": {
      "get": {
        "operationId": "getTranche",
        "summary": "Read a tranche: state, hold, latest decision and plain-language sentences",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tranche"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/tranches/{id}/funding": {
      "post": {
        "operationId": "requestFunding",
        "summary": "Reserve funding (a PayPal authorization hold) for a tranche at the version just read",
        "security": [
          {
            "platformKey": [],
            "requestSignature": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Same key and body return the same response; a different body is 409",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted (recorded, not yet done)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Funding"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid platform key or signature",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or not owned by this platform",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency or version conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "Storage or the feature is unavailable; retry with the same idempotency key",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Allowance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "const": "DRAFT"
          },
          "payee_ref": {
            "type": "string"
          },
          "cap": {
            "$ref": "#/components/schemas/Money"
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          },
          "window_days": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "max_resubmits": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "tranches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrancheRef"
            }
          }
        },
        "required": [
          "id",
          "status",
          "payee_ref",
          "cap",
          "milestones",
          "window_days",
          "max_resubmits",
          "tranches"
        ],
        "additionalProperties": false
      },
      "AllowanceDraft": {
        "type": "object",
        "properties": {
          "payee_ref": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "cap": {
            "$ref": "#/components/schemas/Money"
          },
          "milestones": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          },
          "window_days": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "max_resubmits": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "payee_ref",
          "cap",
          "milestones",
          "window_days",
          "max_resubmits"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "platformKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "The platform key"
      },
      "requestSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Stood-Signature",
        "description": "t=<unix seconds>,v2=<hex HMAC-SHA256(secret, JSON [\"stood.request@2\", t, method, \"/v1\"+path+query, Idempotency-Key, If-Match, Content-Type, body])>"
      }
    }
  }
}